Added "codeClass" option
Code class is a class to add to the `<code>` tag by default.

This can come useful in cases where there is a css that styles the code with syntax highlighting.

For example if we combine marked with [highlight.js](http://highlightjs.org/), a `<pre><code>` block generated from markdown **language** won't have the expected style, that's because marked doesn't add any class to the block, and therefore the css is not applyed. In the same way to acutally highlight a block that **has a language** we would need to use the `langPrefix` option setting it to `"hljs "`, but this is not the usage for a prefix.

One could argue this is a fault in highlight.js, but I would say that it is not. The libray adds the style to the `<pre><code>` blocks that have the defined class, and this is a very good approach that lets the developer the option to use multiple syntax highlighting style, or whatever else they want to do with `<code>`.